[#54] create room 버튼 헤더 이식 및 규격 통일
