+++ frontend/src/stylesheets/style.module.scss
@@ -104,20 +104,10 @@
     
 }
 
-
-.modalContainer{
-    padding-left: 1em;
-    
-    
-}
 .headerUploadButton{
     border: 0;
-    // background-color: white ;
-    // color: #337ab7;
     background: transparent;
     color: $purple;
-    // width: 5.5em;
-    // height: 2em;
     width: 140px;
     height: 120px;
     border-radius: 10px;
